pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentevalkit"
version = "0.6.0"
description = "Testing and evaluation framework for AI agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Amit Paz"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "httpx>=0.24",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
stats = ["scipy>=1.9"]
semantic = ["sentence-transformers>=2.0"]
crewai = ["crewai>=0.28"]
autogen = ["pyautogen>=0.2"]
distributed = [
    "redis>=4.0",
]
dashboard = []  # Currently uses stdlib only
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "build",
    "fakeredis>=2.0",
]

[project.urls]
Homepage = "https://github.com/agentkitai/agenteval"
Repository = "https://github.com/agentkitai/agenteval"

[project.scripts]
agenteval = "agenteval.cli:cli"

[tool.setuptools.package-data]
agenteval = ["dashboard/static/**"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501", "B007", "B904", "SIM108", "SIM105", "SIM103", "UP006", "UP035", "UP045", "B028"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
check_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "agenteval.distributed.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "agenteval.dashboard.*"
ignore_errors = true
